fixes after moving files
