+++ Arrt/View/BlobExplorer/BlobListView.cpp
@@ -5,7 +5,7 @@
 #include <QPainter>
 #include <QPen>
 #include <QStyledItemDelegate>
-#include <Utils/DpiUtils.h>
+#include <ViewUtils/DpiUtils.h>
 #include <View/ArrtStyle.h>
 #include <View/BlobExplorer/BlobListView.h>
 #include <ViewModel/BlobExplorer/BlobsListModel.h>
