+++ ArrtModel/Utils/StyleUtils.h
@@ -1,0 +1,8 @@
+#pragma once
+#include <QString>
+
+namespace StyleUtils
+{
+    // helper function to format parameters in rich text
+    QString formatParameterList(const QStringList& parameterNames);
+} // namespace StyleUtils
